rename bundler.ts to bundle.ts
